# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tidarr API",
    "description": "Tidarr REST API for managing Tidal music downloads, queue management, synchronization, and Lidarr integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8484",
      "description": "Local Tidarr instance"
    }
  ],
  "security": [
    {
      "ApiKeyHeader": []
    },
    {
      "ApiKeyQuery": []
    },
    {
      "BearerToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key passed via X-Api-Key header (recommended for automation)"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "API key passed via query parameter"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth"
      }
    },
    "schemas": {
...
      "CustomCSSResponse": {
        "type": "object",
        "required": [
          "css"
        ],
        "properties": {
          "css": {
            "type": "string"
          }
        }
      },
      "SaveSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SabnzbdVersionResponse": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string"
          }
        }
      },
      "SabnzbdAddResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "nzo_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SabnzbdGetConfigResponse": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "properties": {
              "version": {
                "type": "string"
              },
              "categories": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "priority": {
                      "type": "integer"
                    },
                    "pp": {
                      "type": "string"
                    },
                    "script": {
                      "type": "string"
                    },
                    "dir": {
                      "type": "string"
                    }
                  }
                }
              },
              "misc": {
                "type": "object",
                "properties": {
                  "complete_dir": {
                    "type": "string"
                  },
                  "download_dir": {
                    "type": "string"
                  },
                  "api_key": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "SabnzbdQueueResponse": {
        "type": "object",
        "properties": {
          "queue": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "Downloading",
                  "Idle",
                  "Paused"
                ]
              },
              "paused": {
                "type": "boolean"
              },
              "pause_int": {
                "type": "string"
              },
              "speedlimit": {
                "type": "string"
              },
              "speedlimit_abs": {
                "type": "string"
              },
              "noofslots": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "start": {
                "type": "integer"
              },
              "finish": {
                "type": "integer"
              },
              "slots": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "Downloading",
                        "Queued",
                        "Paused"
                      ]
                    },
                    "index": {
                      "type": "integer"
                    },
                    "eta": {
                      "type": "string"
                    },
                    "timeleft": {
                      "type": "string"
                    },
                    "avg_age": {
                      "type": "string"
                    },
                    "mb": {
                      "type": "string"
                    },
                    "mbleft": {
                      "type": "string"
                    },
                    "mbmissing": {
                      "type": "string"
                    },
                    "size": {
                      "type": "string"
                    },
                    "sizeleft": {
                      "type": "string"
                    },
                    "filename": {
                      "type": "string"
                    },
                    "priority": {
                      "type": "string"
                    },
                    "cat": {
                      "type": "string"
                    },
                    "percentage": {
                      "type": "string"
                    },
                    "nzo_id": {
                      "type": "string"
                    },
                    "unpackopts": {
                      "type": "string"
                    },
                    "labels": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "SabnzbdHistoryResponse": {
        "type": "object",
        "properties": {
          "history": {
            "type": "object",
            "properties": {
              "noofslots": {
                "type": "integer"
              },
              "month_size": {
                "type": "string"
              },
              "week_size": {
                "type": "string"
              },
              "day_size": {
                "type": "string"
              },
              "total_size": {
                "type": "string"
              },
              "slots": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "Completed",
                        "Failed"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "nzo_id": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "size": {
                      "type": "string"
                    },
                    "bytes": {
                      "type": "string"
                    },
                    "fail_message": {
                      "type": "string"
                    },
                    "download_time": {
                      "type": "integer"
                    },
                    "downloaded": {
                      "type": "integer"
                    },
                    "completeness": {
                      "type": "integer"
                    },
                    "script": {
                      "type": "string"
                    },
                    "script_log": {
                      "type": "string"
                    },
                    "script_line": {
                      "type": "string"
                    },
                    "download_name": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "storage": {
                      "type": "string"
                    },
                    "status_string": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/custom-css": {
      "get": {
        "operationId": "getCustomCss",
        "summary": "Get custom CSS",
        "description": "Returns the custom CSS content used to style the Tidarr web UI.",
        "responses": {
          "200": {
            "description": "Custom CSS content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomCSSResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "saveCustomCss",
        "summary": "Save custom CSS",
        "description": "Save custom CSS content to style the Tidarr web UI.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "css"
                ],
                "properties": {
                  "css": {
                    "type": "string",
                    "description": "CSS content"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CSS saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "CSS content must be a string"
          }
        }
      }
    },
...
    "/api/lidarr": {
      "get": {
        "operationId": "lidarrIndexer",
        "summary": "Newznab indexer API for Lidarr",
        "description": "Implements the Newznab protocol for Lidarr integration. Supports capabilities (t=caps), search (t=search), and music search (t=music).",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "caps",
                "search",
                "music"
              ]
            },
            "description": "Request type: caps (capabilities), search (general search), music (artist+album search)"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query (for t=search)"
          },
          {
            "name": "artist",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Artist name (for t=music)"
          },
          {
            "name": "album",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Album name (for t=music)"
          }
        ],
        "responses": {
          "200": {
            "description": "Newznab XML response with capabilities or search results",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/lidarr/download/{id}/{quality}": {
      "get": {
        "operationId": "lidarrDownload",
        "summary": "Download album via Lidarr",
        "description": "Called by Lidarr when grabbing an album. Returns an NZB file that triggers the download. Quality values come from the indexer search results.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Tidal album ID"
          },
          {
            "name": "quality",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "hires_lossless",
                "lossless",
                "high",
                "low"
              ]
            },
            "description": "Download quality as returned by the indexer"
          }
        ],
        "responses": {
          "200": {
            "description": "NZB file",
            "content": {
              "application/x-nzb": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/sabnzbd/api": {
      "get": {
        "operationId": "sabnzbdApiGet",
        "summary": "SABnzbd-compatible download client API (GET)",
        "description": "SABnzbd-compatible API for Lidarr integration. Supports modes: version, get_config, addurl, queue, history.",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "version",
                "get_config",
                "addurl",
                "addfile",
                "queue",
                "history"
              ]
            },
            "description": "SABnzbd API mode"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "URL to add (for mode=addurl) or action name like 'delete' (for mode=queue/history)"
          },
          {
            "name": "value",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "nzo_id to delete (for mode=queue&name=delete or mode=history&name=delete)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Number of history items to return (for mode=history)"
          }
        ],
        "responses": {
          "200": {
            "description": "Response varies by mode",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SabnzbdVersionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SabnzbdGetConfigResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SabnzbdAddResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SabnzbdQueueResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SabnzbdHistoryResponse"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "sabnzbdApiPost",
        "summary": "SABnzbd-compatible download client API (POST)",
        "description": "Supports all SABnzbd modes via POST. Primarily used by Lidarr to upload NZB files (mode=addfile) with multipart/form-data.",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "version",
                "get_config",
                "addurl",
                "addfile",
                "queue",
                "history"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "format": "binary",
                    "description": "NZB file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SabnzbdAddResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}